chore: refactor and add test
