+++ lambda/events/event_create_s3_parameter_raw_simple.json
@@ -1,0 +1,17 @@
+{
+  "RequestType": "Create",
+  "LogicalResourceId": "LogicalResourceId",
+  "ResourceProperties": {
+    "ParameterName": "arn:aws:ssm:eu-central-1:123456789012:parameter/testsecret",
+    "SopsS3File": {
+      "Bucket": "..",
+      "Key": "../test-secrets/binary/sopsfile.enc-age.binary"
+    },
+    "Format": "binary",
+    "ConvertToJSON": "false",
+    "SopsAgeKey": "AGE-SECRET-KEY-1EFUWJ0G2XJTJFWTAM2DGMA4VCK3R05W58FSMHZP3MZQ0ZTAQEAFQC6T7T3"
+  },
+  "ResourceType": "Custom::SOPS::Secret",
+  "RequestId": "RequestId",
+  "StackId": "StackId"
+}
