+++ lambda/events/event_create_s3_secret_env_as_json_simple.json
