+++ lambda/events/event_create_s3_secret_json_complex.json
