+++ lambda/events/event_create_s3_secret_raw_simple.json
