+++ lambda/events/event_create_s3_secret_yaml_as_json_complex_flat.json
