+++ lambda/events/event_create_s3_secret_yaml_complex.json
