+++ lambda/handler_parameter_raw_test.go
@@ -1,0 +1,15 @@
+package main
+
+import (
+	"testing"
+
+	"github.com/gkampitakis/go-snaps/snaps"
+)
+
+func Test_FullWorkflow_Create_S3_Parameter_RAW_Simple(t *testing.T) {
+	mocks, ctx, event := prepareHandler(t, "events/event_create_s3_parameter_raw_simple.json")
+
+	phys, data, err := mocks.syncSopsToSecretsmanager(ctx, event)
+	check(err)
+	snaps.MatchSnapshot(t, ">>>syncSopsToSecretsmanager", phys, data, err)
+}
