+++ lambda/handler_secret_json_test.go
@@ -7,7 +7,7 @@ import (
 )
 
 func Test_FullWorkflow_Create_S3_JSON_Simple(t *testing.T) {
-	mocks, ctx, event := prepareHandler(t, "events/event_create_s3_json_simple.json")
+	mocks, ctx, event := prepareHandler(t, "events/event_create_s3_secret_json_simple.json")
 
 	phys, data, err := mocks.syncSopsToSecretsmanager(ctx, event)
 	check(err)
@@ -15,7 +15,7 @@ func Test_FullWorkflow_Create_S3_JSON_Simple(t *testing.T) {
 }
 
 func Test_FullWorkflow_Create_S3_JSON_Complex(t *testing.T) {
-	mocks, ctx, event := prepareHandler(t, "events/event_create_s3_json_complex.json")
+	mocks, ctx, event := prepareHandler(t, "events/event_create_s3_secret_json_complex.json")
 
 	phys, data, err := mocks.syncSopsToSecretsmanager(ctx, event)
 	check(err)
@@ -23,14 +23,14 @@ func Test_FullWorkflow_Create_S3_JSON_Complex(t *testing.T) {
 }
 
 func Test_FullWorkflow_Create_S3_JSON_Complex_StringifyValues(t *testing.T) {
-	mocks, ctx, event := prepareHandler(t, "events/event_create_s3_json_complex_stringify.json")
+	mocks, ctx, event := prepareHandler(t, "events/event_create_s3_secret_json_complex_stringify.json")
 
 	phys, data, err := mocks.syncSopsToSecretsmanager(ctx, event)
 	check(err)
 	snaps.MatchSnapshot(t, ">>>syncSopsToSecretsmanager", phys, data, err)
 }
 func Test_FullWorkflow_Create_S3_JSON_Complex_Flat(t *testing.T) {
-	mocks, ctx, event := prepareHandler(t, "events/event_create_s3_json_complex_flat.json")
+	mocks, ctx, event := prepareHandler(t, "events/event_create_s3_secret_json_complex_flat.json")
 
 	phys, data, err := mocks.syncSopsToSecretsmanager(ctx, event)
 	check(err)
@@ -38,7 +38,7 @@ func Test_FullWorkflow_Create_S3_JSON_Complex_Flat(t *testing.T) {
 }
 
 func Test_FullWorkflow_Create_INLINE_JSON_Simple(t *testing.T) {
-	mocks, ctx, event := prepareHandler(t, "events/event_create_s3_json_simple.json")
+	mocks, ctx, event := prepareHandler(t, "events/event_create_s3_secret_json_simple.json")
 	event = fileToInline(event)
 
 	phys, data, err := mocks.syncSopsToSecretsmanager(ctx, event)
@@ -47,7 +47,7 @@ func Test_FullWorkflow_Create_INLINE_JSON_Simple(t *testing.T) {
 }
 
 func Test_FullWorkflow_Create_INLINE_JSON_Complex(t *testing.T) {
-	mocks, ctx, event := prepareHandler(t, "events/event_create_s3_json_complex.json")
+	mocks, ctx, event := prepareHandler(t, "events/event_create_s3_secret_json_complex.json")
 	event = fileToInline(event)
 
 	phys, data, err := mocks.syncSopsToSecretsmanager(ctx, event)
@@ -56,7 +56,7 @@ func Test_FullWorkflow_Create_INLINE_JSON_Complex(t *testing.T) {
 }
 
 func Test_FullWorkflow_Create_INLINE_JSON_Complex_StringifyValues(t *testing.T) {
-	mocks, ctx, event := prepareHandler(t, "events/event_create_s3_json_complex_stringify.json")
+	mocks, ctx, event := prepareHandler(t, "events/event_create_s3_secret_json_complex_stringify.json")
 	event = fileToInline(event)
 
 	phys, data, err := mocks.syncSopsToSecretsmanager(ctx, event)
@@ -64,7 +64,7 @@ func Test_FullWorkflow_Create_INLINE_JSON_Complex_StringifyValues(t *testing.T) 
 	snaps.MatchSnapshot(t, ">>>syncSopsToSecretsmanager", phys, data, err)
 }
 func Test_FullWorkflow_Create_INLINE_JSON_Complex_Flat(t *testing.T) {
-	mocks, ctx, event := prepareHandler(t, "events/event_create_s3_json_complex_flat.json")
+	mocks, ctx, event := prepareHandler(t, "events/event_create_s3_secret_json_complex_flat.json")
 	event = fileToInline(event)
 
 	phys, data, err := mocks.syncSopsToSecretsmanager(ctx, event)
