+++ lambda/handler_secret_raw_test.go
@@ -7,7 +7,7 @@ import (
 )
 
 func Test_FullWorkflow_Create_S3_RAW_Simple(t *testing.T) {
-	mocks, ctx, event := prepareHandler(t, "events/event_create_s3_raw_simple.json")
+	mocks, ctx, event := prepareHandler(t, "events/event_create_s3_secret_raw_simple.json")
 
 	phys, data, err := mocks.syncSopsToSecretsmanager(ctx, event)
 	check(err)
