+++ lambda/handler_secret_yaml_test.go
@@ -7,7 +7,7 @@ import (
 )
 
 func Test_FullWorkflow_Create_S3_YAML_Simple(t *testing.T) {
-	mocks, ctx, event := prepareHandler(t, "events/event_create_s3_yaml_simple.json")
+	mocks, ctx, event := prepareHandler(t, "events/event_create_s3_secret_yaml_simple.json")
 
 	phys, data, err := mocks.syncSopsToSecretsmanager(ctx, event)
 	check(err)
@@ -15,7 +15,7 @@ func Test_FullWorkflow_Create_S3_YAML_Simple(t *testing.T) {
 }
 
 func Test_FullWorkflow_Create_S3_YAML_as_JSON_Simple(t *testing.T) {
-	mocks, ctx, event := prepareHandler(t, "events/event_create_s3_yaml_as_json_simple.json")
+	mocks, ctx, event := prepareHandler(t, "events/event_create_s3_secret_yaml_as_json_simple.json")
 
 	phys, data, err := mocks.syncSopsToSecretsmanager(ctx, event)
 	check(err)
@@ -23,7 +23,7 @@ func Test_FullWorkflow_Create_S3_YAML_as_JSON_Simple(t *testing.T) {
 }
 
 func Test_FullWorkflow_Create_S3_YAML_Complex(t *testing.T) {
-	mocks, ctx, event := prepareHandler(t, "events/event_create_s3_yaml_complex.json")
+	mocks, ctx, event := prepareHandler(t, "events/event_create_s3_secret_yaml_complex.json")
 
 	phys, data, err := mocks.syncSopsToSecretsmanager(ctx, event)
 	check(err)
@@ -31,7 +31,7 @@ func Test_FullWorkflow_Create_S3_YAML_Complex(t *testing.T) {
 }
 
 func Test_FullWorkflow_Create_S3_YAML_as_JSON_Complex(t *testing.T) {
-	mocks, ctx, event := prepareHandler(t, "events/event_create_s3_yaml_as_json_complex.json")
+	mocks, ctx, event := prepareHandler(t, "events/event_create_s3_secret_yaml_as_json_complex.json")
 
 	phys, data, err := mocks.syncSopsToSecretsmanager(ctx, event)
 	check(err)
@@ -39,7 +39,7 @@ func Test_FullWorkflow_Create_S3_YAML_as_JSON_Complex(t *testing.T) {
 }
 
 func Test_FullWorkflow_Create_S3_YAML_Complex_Flat(t *testing.T) {
-	mocks, ctx, event := prepareHandler(t, "events/event_create_s3_yaml_complex_flat.json")
+	mocks, ctx, event := prepareHandler(t, "events/event_create_s3_secret_yaml_complex_flat.json")
 
 	phys, data, err := mocks.syncSopsToSecretsmanager(ctx, event)
 	check(err)
@@ -47,7 +47,7 @@ func Test_FullWorkflow_Create_S3_YAML_Complex_Flat(t *testing.T) {
 }
 
 func Test_FullWorkflow_Create_S3_YAML_as_JSON_Complex_Flat(t *testing.T) {
-	mocks, ctx, event := prepareHandler(t, "events/event_create_s3_yaml_as_json_complex_flat.json")
+	mocks, ctx, event := prepareHandler(t, "events/event_create_s3_secret_yaml_as_json_complex_flat.json")
 
 	phys, data, err := mocks.syncSopsToSecretsmanager(ctx, event)
 	check(err)
@@ -55,7 +55,7 @@ func Test_FullWorkflow_Create_S3_YAML_as_JSON_Complex_Flat(t *testing.T) {
 }
 
 func Test_FullWorkflow_Create_INLINE_YAML_Simple(t *testing.T) {
-	mocks, ctx, event := prepareHandler(t, "events/event_create_s3_yaml_simple.json")
+	mocks, ctx, event := prepareHandler(t, "events/event_create_s3_secret_yaml_simple.json")
 	event = fileToInline(event)
 
 	phys, data, err := mocks.syncSopsToSecretsmanager(ctx, event)
@@ -64,7 +64,7 @@ func Test_FullWorkflow_Create_INLINE_YAML_Simple(t *testing.T) {
 }
 
 func Test_FullWorkflow_Create_INLINE_YAML_as_JSON_Simple(t *testing.T) {
-	mocks, ctx, event := prepareHandler(t, "events/event_create_s3_yaml_as_json_simple.json")
+	mocks, ctx, event := prepareHandler(t, "events/event_create_s3_secret_yaml_as_json_simple.json")
 	event = fileToInline(event)
 
 	phys, data, err := mocks.syncSopsToSecretsmanager(ctx, event)
@@ -73,7 +73,7 @@ func Test_FullWorkflow_Create_INLINE_YAML_as_JSON_Simple(t *testing.T) {
 }
 
 func Test_FullWorkflow_Create_INLINE_YAML_Complex(t *testing.T) {
-	mocks, ctx, event := prepareHandler(t, "events/event_create_s3_yaml_complex.json")
+	mocks, ctx, event := prepareHandler(t, "events/event_create_s3_secret_yaml_complex.json")
 	event = fileToInline(event)
 
 	phys, data, err := mocks.syncSopsToSecretsmanager(ctx, event)
@@ -82,7 +82,7 @@ func Test_FullWorkflow_Create_INLINE_YAML_Complex(t *testing.T) {
 }
 
 func Test_FullWorkflow_Create_INLINE_YAML_as_JSON_Complex(t *testing.T) {
-	mocks, ctx, event := prepareHandler(t, "events/event_create_s3_yaml_as_json_complex.json")
+	mocks, ctx, event := prepareHandler(t, "events/event_create_s3_secret_yaml_as_json_complex.json")
 
 	phys, data, err := mocks.syncSopsToSecretsmanager(ctx, event)
 	check(err)
@@ -90,7 +90,7 @@ func Test_FullWorkflow_Create_INLINE_YAML_as_JSON_Complex(t *testing.T) {
 }
 
 func Test_FullWorkflow_Create_INLINE_YAML_Complex_Flat(t *testing.T) {
-	mocks, ctx, event := prepareHandler(t, "events/event_create_s3_yaml_complex_flat.json")
+	mocks, ctx, event := prepareHandler(t, "events/event_create_s3_secret_yaml_complex_flat.json")
 	event = fileToInline(event)
 
 	phys, data, err := mocks.syncSopsToSecretsmanager(ctx, event)
@@ -99,7 +99,7 @@ func Test_FullWorkflow_Create_INLINE_YAML_Complex_Flat(t *testing.T) {
 }
 
 func Test_FullWorkflow_Create_INLINE_YAML_as_JSON_Complex_Flat(t *testing.T) {
-	mocks, ctx, event := prepareHandler(t, "events/event_create_s3_yaml_as_json_complex_flat.json")
+	mocks, ctx, event := prepareHandler(t, "events/event_create_s3_secret_yaml_as_json_complex_flat.json")
 	event = fileToInline(event)
 
 	phys, data, err := mocks.syncSopsToSecretsmanager(ctx, event)
