+++ test/secret-asset.integ.snapshot/SecretIntegrationAsset.assets.json
@@ -1,15 +1,15 @@
 {
   "version": "36.0.0",
   "files": {
-    "85bc7001ed80d8e9d813ad236a8baa851799b8fa751e55df20b6052b6971fdaa": {
+    "092d58cc76fd73ef78f112106586391ec312b04bc9e46b4a46b9199c2a78dec4": {
       "source": {
-        "path": "asset.85bc7001ed80d8e9d813ad236a8baa851799b8fa751e55df20b6052b6971fdaa.zip",
+        "path": "asset.092d58cc76fd73ef78f112106586391ec312b04bc9e46b4a46b9199c2a78dec4.zip",
         "packaging": "file"
       },
       "destinations": {
         "current_account-current_region": {
           "bucketName": "cdk-hnb659fds-assets-${AWS::AccountId}-${AWS::Region}",
-          "objectKey": "85bc7001ed80d8e9d813ad236a8baa851799b8fa751e55df20b6052b6971fdaa.zip",
+          "objectKey": "092d58cc76fd73ef78f112106586391ec312b04bc9e46b4a46b9199c2a78dec4.zip",
           "assumeRoleArn": "arn:${AWS::Partition}:iam::${AWS::AccountId}:role/cdk-hnb659fds-file-publishing-role-${AWS::AccountId}-${AWS::Region}"
         }
       }
@@ -79,7 +79,7 @@
         }
       }
     },
-    "22ad8b3d36812c0748d2899eb0d8fbd044c5104b025f7c21f79c1302201fefc5": {
+    "d2954cb57c9c33fc2eb3542d6495649d78639813f6faa7822ce0f1f8c3bae3cd": {
       "source": {
         "path": "SecretIntegrationAsset.template.json",
         "packaging": "file"
@@ -87,7 +87,7 @@
       "destinations": {
         "current_account-current_region": {
           "bucketName": "cdk-hnb659fds-assets-${AWS::AccountId}-${AWS::Region}",
-          "objectKey": "22ad8b3d36812c0748d2899eb0d8fbd044c5104b025f7c21f79c1302201fefc5.json",
+          "objectKey": "d2954cb57c9c33fc2eb3542d6495649d78639813f6faa7822ce0f1f8c3bae3cd.json",
           "assumeRoleArn": "arn:${AWS::Partition}:iam::${AWS::AccountId}:role/cdk-hnb659fds-file-publishing-role-${AWS::AccountId}-${AWS::Region}"
         }
       }
