+++ test/secret-asset.integ.snapshot/SecretIntegrationAsset.template.json
@@ -229,7 +229,7 @@
      "S3Bucket": {
       "Fn::Sub": "cdk-hnb659fds-assets-${AWS::AccountId}-${AWS::Region}"
      },
-     "S3Key": "85bc7001ed80d8e9d813ad236a8baa851799b8fa751e55df20b6052b6971fdaa.zip"
+     "S3Key": "092d58cc76fd73ef78f112106586391ec312b04bc9e46b4a46b9199c2a78dec4.zip"
     },
     "Environment": {
      "Variables": {
