+++ test/secret-inline.integ.snapshot/SecretIntegrationInline.assets.json
@@ -1,20 +1,20 @@
 {
   "version": "36.0.0",
   "files": {
-    "85bc7001ed80d8e9d813ad236a8baa851799b8fa751e55df20b6052b6971fdaa": {
+    "092d58cc76fd73ef78f112106586391ec312b04bc9e46b4a46b9199c2a78dec4": {
       "source": {
-        "path": "asset.85bc7001ed80d8e9d813ad236a8baa851799b8fa751e55df20b6052b6971fdaa.zip",
+        "path": "asset.092d58cc76fd73ef78f112106586391ec312b04bc9e46b4a46b9199c2a78dec4.zip",
         "packaging": "file"
       },
       "destinations": {
         "current_account-current_region": {
           "bucketName": "cdk-hnb659fds-assets-${AWS::AccountId}-${AWS::Region}",
-          "objectKey": "85bc7001ed80d8e9d813ad236a8baa851799b8fa751e55df20b6052b6971fdaa.zip",
+          "objectKey": "092d58cc76fd73ef78f112106586391ec312b04bc9e46b4a46b9199c2a78dec4.zip",
           "assumeRoleArn": "arn:${AWS::Partition}:iam::${AWS::AccountId}:role/cdk-hnb659fds-file-publishing-role-${AWS::AccountId}-${AWS::Region}"
         }
       }
     },
-    "c066b6cf6f68690b807d157895df3ac6371a7fdeebb114179be19e9ac5e4c168": {
+    "967d495ade78ab7c794f4de692504975a0138eec98103d136bcd8e2df420aa1a": {
       "source": {
         "path": "SecretIntegrationInline.template.json",
         "packaging": "file"
@@ -22,7 +22,7 @@
       "destinations": {
         "current_account-current_region": {
           "bucketName": "cdk-hnb659fds-assets-${AWS::AccountId}-${AWS::Region}",
-          "objectKey": "c066b6cf6f68690b807d157895df3ac6371a7fdeebb114179be19e9ac5e4c168.json",
+          "objectKey": "967d495ade78ab7c794f4de692504975a0138eec98103d136bcd8e2df420aa1a.json",
           "assumeRoleArn": "arn:${AWS::Partition}:iam::${AWS::AccountId}:role/cdk-hnb659fds-file-publishing-role-${AWS::AccountId}-${AWS::Region}"
         }
       }
