+++ test/secret-manual.integ.snapshot/SecretIntegrationAsset.assets.json
@@ -1,20 +1,20 @@
 {
   "version": "36.0.0",
   "files": {
-    "85bc7001ed80d8e9d813ad236a8baa851799b8fa751e55df20b6052b6971fdaa": {
+    "092d58cc76fd73ef78f112106586391ec312b04bc9e46b4a46b9199c2a78dec4": {
       "source": {
-        "path": "asset.85bc7001ed80d8e9d813ad236a8baa851799b8fa751e55df20b6052b6971fdaa.zip",
+        "path": "asset.092d58cc76fd73ef78f112106586391ec312b04bc9e46b4a46b9199c2a78dec4.zip",
         "packaging": "file"
       },
       "destinations": {
         "current_account-current_region": {
           "bucketName": "cdk-hnb659fds-assets-${AWS::AccountId}-${AWS::Region}",
-          "objectKey": "85bc7001ed80d8e9d813ad236a8baa851799b8fa751e55df20b6052b6971fdaa.zip",
+          "objectKey": "092d58cc76fd73ef78f112106586391ec312b04bc9e46b4a46b9199c2a78dec4.zip",
           "assumeRoleArn": "arn:${AWS::Partition}:iam::${AWS::AccountId}:role/cdk-hnb659fds-file-publishing-role-${AWS::AccountId}-${AWS::Region}"
         }
       }
     },
-    "e6a59ee63ab91485e14ad39235062bcb08d47e7624f0a5e585a78173dfa32139": {
+    "f1cc60e107ed441ecfaef66185bc401f153aabc890305c5ae45a9d0e80c3ce73": {
       "source": {
         "path": "SecretIntegrationAsset.template.json",
         "packaging": "file"
@@ -22,7 +22,7 @@
       "destinations": {
         "current_account-current_region": {
           "bucketName": "cdk-hnb659fds-assets-${AWS::AccountId}-${AWS::Region}",
-          "objectKey": "e6a59ee63ab91485e14ad39235062bcb08d47e7624f0a5e585a78173dfa32139.json",
+          "objectKey": "f1cc60e107ed441ecfaef66185bc401f153aabc890305c5ae45a9d0e80c3ce73.json",
           "assumeRoleArn": "arn:${AWS::Partition}:iam::${AWS::AccountId}:role/cdk-hnb659fds-file-publishing-role-${AWS::AccountId}-${AWS::Region}"
         }
       }
