+++ test/secret-multikms.integ.snapshot/SecretMultiKms.assets.json
@@ -1,20 +1,20 @@
 {
   "version": "36.0.0",
   "files": {
-    "85bc7001ed80d8e9d813ad236a8baa851799b8fa751e55df20b6052b6971fdaa": {
+    "092d58cc76fd73ef78f112106586391ec312b04bc9e46b4a46b9199c2a78dec4": {
       "source": {
-        "path": "asset.85bc7001ed80d8e9d813ad236a8baa851799b8fa751e55df20b6052b6971fdaa.zip",
+        "path": "asset.092d58cc76fd73ef78f112106586391ec312b04bc9e46b4a46b9199c2a78dec4.zip",
         "packaging": "file"
       },
       "destinations": {
         "current_account-current_region": {
           "bucketName": "cdk-hnb659fds-assets-${AWS::AccountId}-${AWS::Region}",
-          "objectKey": "85bc7001ed80d8e9d813ad236a8baa851799b8fa751e55df20b6052b6971fdaa.zip",
+          "objectKey": "092d58cc76fd73ef78f112106586391ec312b04bc9e46b4a46b9199c2a78dec4.zip",
           "assumeRoleArn": "arn:${AWS::Partition}:iam::${AWS::AccountId}:role/cdk-hnb659fds-file-publishing-role-${AWS::AccountId}-${AWS::Region}"
         }
       }
     },
-    "728cf0fb4de71994fe1173c693b75610ff3c6c6f45b7f5f7db0df9a7d07dc0e5": {
+    "4c795b4eafe6fc449c9c07c0cd791648009d1051cc327266c404c758580c9c16": {
       "source": {
         "path": "SecretMultiKms.template.json",
         "packaging": "file"
@@ -22,7 +22,7 @@
       "destinations": {
         "current_account-current_region": {
           "bucketName": "cdk-hnb659fds-assets-${AWS::AccountId}-${AWS::Region}",
-          "objectKey": "728cf0fb4de71994fe1173c693b75610ff3c6c6f45b7f5f7db0df9a7d07dc0e5.json",
+          "objectKey": "4c795b4eafe6fc449c9c07c0cd791648009d1051cc327266c404c758580c9c16.json",
           "assumeRoleArn": "arn:${AWS::Partition}:iam::${AWS::AccountId}:role/cdk-hnb659fds-file-publishing-role-${AWS::AccountId}-${AWS::Region}"
         }
       }
